+++ halo-third-party-sdk-model/CHANGELOG.rst
@@ -2,6 +2,12 @@
 CHANGELOG
 =========
 
+1.1.0
+-----
+* Activity data object was updated to include Activity Summary, and Activity Sessions now support intensity durations.
+* Sleep data object was updated to include session start and end times.
+* Notification type was updated to include 'HEALTH_METRIC_DELETION' and 'HEALTH_METRIC_DELETION_TEST'.
+
 1.0.0
 -----
 * Initial release of Halo Third Party SDK.